KS-Backend: Rename employer specific classes/files
Rename:
 - ApplicationExcelDownloadView to EmployerApplicationExcelDownloadView
 - ApplicationFactory to EmployerApplicationFactory
 - ApplicationStatusValidator to EmployerApplicationStatusValidator
 - clean_old_applications.py to clean_old_employer_applications.py

Also remove import renamings from test_cron_job.py.

Refs YJDH-493 (Refactoring before adding YouthApplicationStatus)
